fixes table name in create_user.sql that would fail in case-sensitive databases
The OAC_ResourceExternalID table was spelled with different case in the create_user.sql script, which
would cause no privileges to be granted on that table within a case-sensitive database environment.
Reverts the table name back to the original (and ensures it is consistent across all scripts) because
we decided to modify the corresponding SQL statements in oacc-core to match the correct table names
exactly, rather than regenerate *all* database scripts with corrected names for consistency.